client class moved to a new file and then injected into ATM file
