+++ lib/bank.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'date'
+require 'client'
 
 class ATM
 
@@ -27,22 +28,3 @@ class ATM
   end
 end
 
-class Client
-  attr_accessor :balance, :log
-  def initialize
-    @balance = 0
-    @log = []
-  end
-
-  def deposit(amount)
-    @balance += amount
-    @log << [Time.now.strftime('%d/%m/%Y'), amount.to_s + '.00', '',
-             @balance.to_s + '.00']
-  end
-
-  def withdraw(amount)
-    @balance -= amount
-    @log << [Time.now.strftime('%d/%m/%Y'), '', amount.to_s + '.00',
-             balance.to_s + '.00']
-  end
-end
